refactor: rename the 'out.persistence.jpa' pathname e classes to 'persistence'
