+++ src/main/java/br/com/fiap/grupo30/fastfood/application/mapper/impl/CustomerDTOMapper.java
@@ -2,7 +2,7 @@ package br.com.fiap.grupo30.fastfood.application.mapper.impl;
 
 import br.com.fiap.grupo30.fastfood.application.dto.CustomerDTO;
 import br.com.fiap.grupo30.fastfood.application.mapper.Mapper;
-import br.com.fiap.grupo30.fastfood.infrastructure.out.persistence.jpa.entities.CustomerEntity;
+import br.com.fiap.grupo30.fastfood.infrastructure.persistence.entities.CustomerEntity;
 import org.springframework.stereotype.Component;
 
 @Component
