+++ src/main/java/br/com/fiap/grupo30/fastfood/domain/services/impl/CategoryServiceImpl.java
@@ -3,7 +3,7 @@ package br.com.fiap.grupo30.fastfood.domain.services.impl;
 import br.com.fiap.grupo30.fastfood.application.dto.CategoryDTO;
 import br.com.fiap.grupo30.fastfood.application.mapper.impl.CategoryMapper;
 import br.com.fiap.grupo30.fastfood.domain.services.CategoryService;
-import br.com.fiap.grupo30.fastfood.infrastructure.out.persistence.jpa.repositories.CategoryRepository;
+import br.com.fiap.grupo30.fastfood.infrastructure.persistence.repositories.CategoryRepository;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
