+++ src/main/java/br/com/fiap/grupo30/fastfood/domain/services/impl/CustomerServiceImpl.java
@@ -8,8 +8,8 @@ import br.com.fiap.grupo30.fastfood.application.mapper.impl.CustomerDTOMapper;
 import br.com.fiap.grupo30.fastfood.application.mapper.impl.CustomerMapper;
 import br.com.fiap.grupo30.fastfood.domain.services.CustomerService;
 import br.com.fiap.grupo30.fastfood.domain.vo.CPF;
-import br.com.fiap.grupo30.fastfood.infrastructure.out.persistence.jpa.entities.CustomerEntity;
-import br.com.fiap.grupo30.fastfood.infrastructure.out.persistence.jpa.repositories.CustomerRepository;
+import br.com.fiap.grupo30.fastfood.infrastructure.persistence.entities.CustomerEntity;
+import br.com.fiap.grupo30.fastfood.infrastructure.persistence.repositories.CustomerRepository;
 import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
