+++ src/main/java/br/com/fiap/grupo30/fastfood/infrastructure/persistence/repositories/CategoryRepository.java
@@ -1,6 +1,6 @@
-package br.com.fiap.grupo30.fastfood.infrastructure.out.persistence.jpa.repositories;
+package br.com.fiap.grupo30.fastfood.infrastructure.persistence.repositories;
 
-import br.com.fiap.grupo30.fastfood.infrastructure.out.persistence.jpa.entities.CategoryEntity;
+import br.com.fiap.grupo30.fastfood.infrastructure.persistence.entities.CategoryEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
