+++ src/main/java/br/com/fiap/grupo30/fastfood/infrastructure/persistence/repositories/CustomerRepository.java
@@ -1,6 +1,6 @@
-package br.com.fiap.grupo30.fastfood.infrastructure.out.persistence.jpa.repositories;
+package br.com.fiap.grupo30.fastfood.infrastructure.persistence.repositories;
 
-import br.com.fiap.grupo30.fastfood.infrastructure.out.persistence.jpa.entities.CustomerEntity;
+import br.com.fiap.grupo30.fastfood.infrastructure.persistence.entities.CustomerEntity;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
